+++ cmd/substreams/init.go
@@ -123,11 +123,6 @@ func runSubstreamsInitE(cmd *cobra.Command, args []string) error {
 			}
 		}
 
-		sinkChoice, err := promptSinkChoice()
-		if err != nil {
-			return fmt.Errorf("running sink choice prompt")
-		}
-
 		fmt.Printf("Retrieving %s contract information (ABI & creation block)\n", chain.DisplayName)
 
 		// Get contract abiContents & parse them
@@ -159,7 +154,6 @@ func runSubstreamsInitE(cmd *cobra.Command, args []string) error {
 			chain,
 			ethereumContracts,
 			lowestStartBlock,
-			sinkChoice,
 		)
 		if err != nil {
 			return fmt.Errorf("new Ethereum %s project: %w", chain.DisplayName, err)
@@ -184,6 +178,14 @@ func runSubstreamsInitE(cmd *cobra.Command, args []string) error {
 	}
 
 	fmt.Printf("Project %q initialized at %q\n", projectName, absoluteWorkingDir)
+	fmt.Println()
+	fmt.Println("Run 'make build' to build the wasm code.")
+	fmt.Println()
+	fmt.Println("The following substreams.yaml files have been created with different sink targets:")
+	fmt.Println(" * substreams.yaml: no sink target")
+	fmt.Println(" * substreams.sql.yaml: PostgreSQL sink")
+	fmt.Println(" * substreams.clickhouse.yaml: Clickhouse sink")
+	fmt.Println(" * substreams.subgraph.yaml: Sink into Substreams-based subgraph")
 
 	return nil
 }
@@ -341,41 +343,6 @@ func promptEthereumContractShortNames(ethereumContracts []*templates.EthereumCon
 	}
 
 	return ethereumContracts, nil
-}
-
-func promptSinkChoice() (codegen.SinkChoice, error) {
-	if devInitSinkChoice != "" {
-		sinkChoice, err := codegen.ParseSinkChoice(devInitSinkChoice)
-		if err != nil {
-			return 0, fmt.Errorf("invalid sink choice %s: %w", sinkChoice, err)
-		}
-	}
-
-	choice := promptui.Select{
-		Label: "Select Sink choice",
-		Items: codegen.SinkChoiceNames(),
-		Templates: &promptui.SelectTemplates{
-			Selected: `{{ "Sink:" | faint }} {{ . }}`,
-		},
-		HideHelp: true,
-	}
-
-	_, selection, err := choice.Run()
-	if err != nil {
-		if errors.Is(err, promptui.ErrInterrupt) {
-			// We received Ctrl-C, users wants to abort, nothing else to do, quit immediately
-			os.Exit(1)
-		}
-
-		return 0, fmt.Errorf("running protocol prompt: %w", err)
-	}
-
-	var sinkChoice codegen.SinkChoice
-	if err := sinkChoice.UnmarshalText([]byte(selection)); err != nil {
-		panic(fmt.Errorf("impossible, selecting hard-coded value from enum itself, something is really wrong here"))
-	}
-
-	return sinkChoice, nil
 }
 
 func promptTrackContract() (bool, error) {