+++ codegen/templates/ethereum/schema.clickhouse.sql
@@ -1,0 +1,36 @@
+CREATE TABLE IF NOT EXISTS approval (
+    "evt_tx_hash" VARCHAR(64),
+    "evt_index" INT,
+    "evt_block_time" TIMESTAMP,
+    "evt_block_number" UInt64,
+    "approved" VARCHAR(40),
+    "owner" VARCHAR(40),
+    "token_id" UInt32
+) ENGINE = MergeTree PRIMARY KEY ("evt_tx_hash","evt_index");
+CREATE TABLE IF NOT EXISTS approval_for_all (
+    "evt_tx_hash" VARCHAR(64),
+    "evt_index" INT,
+    "evt_block_time" TIMESTAMP,
+    "evt_block_number" UInt64,
+    "approved" BOOL,
+    "operator" VARCHAR(40),
+    "owner" VARCHAR(40)
+) ENGINE = MergeTree PRIMARY KEY ("evt_tx_hash","evt_index");
+CREATE TABLE IF NOT EXISTS ownership_transferred (
+    "evt_tx_hash" VARCHAR(64),
+    "evt_index" INT,
+    "evt_block_time" TIMESTAMP,
+    "evt_block_number" UInt64,
+    "new_owner" VARCHAR(40),
+    "previous_owner" VARCHAR(40)
+) ENGINE = MergeTree PRIMARY KEY ("evt_tx_hash","evt_index");
+CREATE TABLE IF NOT EXISTS transfer (
+    "evt_tx_hash" VARCHAR(64),
+    "evt_index" INT,
+    "evt_block_time" TIMESTAMP,
+    "evt_block_number" UInt64,
+    "from" VARCHAR(40),
+    "to" VARCHAR(40),
+    "token_id" UInt32
+) ENGINE = MergeTree PRIMARY KEY ("evt_tx_hash","evt_index");
+